Fix mismatch of calibration-function signatures
- The member function of EnergySensitiveDetectorChannel expected the event number (integer) first, and the channel (double) second. This was in contradiction to the IO functions of carolina and led to rounding errors.

+++ include/detectors/energy_sensitive_detector_channel.hpp
@@ -53,7 +53,7 @@ struct EnergySensitiveDetectorChannel final : public Channel {
                 return true;
             });
 
-    const function<double(const int, const double)> energy_calibration;
+    const function<double(const double, const long long)> energy_calibration;
     const function<double(const double, const double)> time_calibration;
     const function<bool(const double)> time_vs_reference_time_gate;
